requirements.txt
black==22.12.0
certifi==2020.12.5
chardet==3.0.4
click==8.1.3
filestack-python==3.2.1
fpdf==1.7.2
idna==2.10
mypy-extensions==0.4.3
numpy==1.24.0
pathspec==0.10.3
Pillow==9.3.0
platformdirs==2.6.0
requests==2.23.0
tomli==2.0.1
trafaret==2.0.2
urllib3==1.25.11
